Moves shared search form functionality to a mixin.
