+++ app/forms/search_form.rb
@@ -1,23 +1,9 @@
 class SearchForm
   include ActiveModel::Model
   include ActiveModel::Validations::Callbacks
+  include SearchFormFilters
 
-  TYPES_OF_ADVICE = [
-    :options_when_paying_for_care,
-    :equity_release,
-    :inheritance_tax_planning,
-    :wills_and_probate
-  ]
-
-  ANY_SIZE_VALUE = 'any'
-
-  attr_accessor :postcode,
-    :checkbox,
-    :coordinates,
-    :pension_transfer,
-    :pension_pot,
-    :pension_pot_size,
-    *TYPES_OF_ADVICE
+  attr_accessor :postcode, :coordinates
 
   before_validation :upcase_postcode
 
@@ -29,24 +15,6 @@ class SearchForm
 
   def coordinates
     @coordinates ||= Geocode.call(postcode)
-  end
-
-  def types_of_advice
-    TYPES_OF_ADVICE.select { |type| public_send(type) == '1' }
-  end
-
-  def pension_pot?
-    pension_pot == '1'
-  end
-
-  def pension_pot_sizes
-    InvestmentSize.all.map do |investment_size|
-      [investment_size.localized_name, investment_size.id]
-    end << [I18n.t('search_filter.pension_pot.any_size_option'), ANY_SIZE_VALUE]
-  end
-
-  def any_pension_pot_size?
-    pension_pot_size && pension_pot_size == ANY_SIZE_VALUE
   end
 
   private
